refactor(observer): colocate session recovery into a named module
Recovery resolve, inventory, assessment, eligibility, and selection were sibling roots; grouping them names the responsibility without changing resume behavior.
