+++ apps/observer/src/commands/session/resumeAgent.ts
@@ -6,7 +6,7 @@ import type { EventJournal, SessionStore } from "../../persistence/index.js";
 import type { ProviderRegistry } from "../../providers/registry.js";
 import type { ObserverCore } from "../../reconcile/core.js";
 import type { ObserverEventBus } from "../../runtime/eventBus.js";
-import { resolveSessionRecovery } from "../../sessionRecovery.js";
+import { resolveSessionRecovery } from "../../sessionRecovery/resolve.js";
 import type { StationLogger } from "../../stationLogger.js";
 import { nowIso } from "../../utils/time.js";
 import {
