+++ apps/observer/src/reconcile/sessionActivation.ts
