+++ apps/observer/src/sessionRecovery/eligibility.ts
