+++ docs/observer-architecture.md
@@ -231,8 +231,7 @@ areas contain the following responsibilities:
 | Area | Current responsibility | Adopted ownership |
 | --- | --- | --- |
 | `commands/` | command queue, routing, scopes, cancellation, launch preflight, direct terminal operations, Group mutation, and command use cases | Driving application behavior; command handlers coordinate launch preflight and provider-neutral terminal operations directly through their narrow ports, while Group mutation remains a dedicated use case. |
-| `sessionRecoveryInventory.ts` | read-only recovery inventory over retained sessions and recovery handles | Application use case that projects provider-neutral, redacted point-in-time evidence without dispatching a command, reconciling providers, or mutating persistence. |
-| `sessionRecoveryAssessment.ts` | read-only recovery classification and newest-eligible handle selection | Application use case that combines one inventory with one captured graph and provider capabilities, reuses the canonical eligibility policy, and returns provider-neutral decisions without launching, reconciling, or mutating persistence. |
+| `sessionRecovery/` | session recovery resolution, inventory, assessment, eligibility, and selection | Application use cases (`resolve.ts`, `inventory.ts`, `assessment.ts`) turn durable session state into typed resume authority or provider-neutral, redacted evidence without launching, reconciling, or mutating persistence; deterministic policies (`eligibility.ts`, `selection.ts`) admit and order recovery handles. |
 | `reconcile/` | provider reads, correlation, graph construction, Group projection, and core state | Reconcile-owned Group repair, command-local Group projection, and deterministic policies; provider I/O remains at its driven edges. `reconcileResult.ts` owns the `ReconcileTiming` result record returned by `runReconcileOnce`, while `core.ts` re-exports it for compatibility. |
 | `hooks/` | raw hook persistence and adapter handoff, report ingestion, dedupe, readiness, spool I/O, and ingress queue | One adapter-to-report normalization path; non-report hooks are reconcile hints, and queue orchestration stays separate from filesystem spool adapters. |
 | `runtime/` | API assembly, process lifecycle, exact read-only ownership inspection, scheduling, event delivery, server bridge, and external launch | Observer composition plus application operations; transport and infrastructure stay at the edge. Exact inspection consumes purpose-specific status, pidfile, cooperative process, and pinned recovery ports without lifecycle mutation authority. |
